Introduce common Input type
