+++ input/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "input"
+version = "0.1.0"
+authors = ["Lucas Soriano del Pino <l.soriano.del.pino@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+elements = "0.16"
+serde = { version = "1", features = [ "derive" ] }
